Struct libgit2-sys::git_revparse_mode_t

pub struct git_revparse_mode_t {
    // some fields omitted
}

Methods

impl git_revparse_mode_t

fn empty() -> git_revparse_mode_t

Returns an empty set of flags.

fn all() -> git_revparse_mode_t

Returns the set containing all flags.

fn bits(&self) -> c_uint

Returns the raw value of the flags currently stored.

fn from_bits(bits: c_uint) -> Option<git_revparse_mode_t>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: c_uint) -> git_revparse_mode_t

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: git_revparse_mode_t) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: git_revparse_mode_t) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: git_revparse_mode_t)

Inserts the specified flags in-place.

fn remove(&mut self, other: git_revparse_mode_t)

Removes the specified flags in-place.

fn toggle(&mut self, other: git_revparse_mode_t)

Toggles the specified flags in-place.